+++ CHANGELOG.md
@@ -8,6 +8,7 @@ User-visible changes worth mentioning.
 ## main
 
 Add your entry here.
+- [#1775] Fix Applications Secret Not Null Constraint generator
 
 ## 5.8.2
 
@@ -31,7 +32,7 @@ Add your entry here.
 - [#1712] Add `Pragma: no-cache` to token response
 - [#1726] Refactor token introspection class.
 - [#1727] Allow to set null secret value for Applications if they are public.
-- [#1735] Add `pkce_code_challenge_methods` config option. 
+- [#1735] Add `pkce_code_challenge_methods` config option.
 
 ## 5.7.1